Refactor handshake validation and optimize utilities
Refactored `IsValid` method from `HandshakeX25519` to `Handshake` for centralized validation logic. Replaced `IsAllZero` with a high-performance `IsZero` method in `BitwiseOperations`. Updated `HandleClientHello` and `OnMessage` to use the new methods.

Adjusted namespaces for consistency, removing unused imports and improving project structure. Added `cancellationToken` support to `ConnectAsync` for better cancellation handling. Improved `NetworkClientService` disconnection handling and clarified `SessionToken` auto-fill logic.

Bumped version to 12.0.5 to reflect these changes.
